FACES-2320 Move the com.liferay.faces.util.render.HttpHeaders interface to package com.liferay.faces.util
