+++ alloy/src/main/java/com/liferay/faces/alloy/component/inputfile/internal/InputFileDecoderPartImpl.java
@@ -26,13 +26,13 @@ import javax.faces.context.FacesContext;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.Part;
 
+import com.liferay.faces.util.HttpHeaders;
 import com.liferay.faces.util.factory.FactoryExtensionFinder;
 import com.liferay.faces.util.lang.StringPool;
 import com.liferay.faces.util.logging.Logger;
 import com.liferay.faces.util.logging.LoggerFactory;
 import com.liferay.faces.util.model.UploadedFile;
 import com.liferay.faces.util.model.UploadedFileFactory;
-import com.liferay.faces.util.render.HttpHeaders;
 
 
 /**
