+++ bridge-impl/src/main/java/com/liferay/faces/bridge/application/internal/ResourceImpl.java
@@ -34,12 +34,12 @@ import javax.faces.context.FacesContext;
 import com.liferay.faces.bridge.config.BridgeConfig;
 import com.liferay.faces.bridge.config.internal.BridgeConfigAttributeMap;
 import com.liferay.faces.bridge.context.BridgeContext;
+import com.liferay.faces.util.HttpHeaders;
 import com.liferay.faces.util.application.ResourceConstants;
 import com.liferay.faces.util.config.ConfiguredServletMapping;
 import com.liferay.faces.util.lang.StringPool;
 import com.liferay.faces.util.logging.Logger;
 import com.liferay.faces.util.logging.LoggerFactory;
-import com.liferay.faces.util.render.HttpHeaders;
 
 
 /**
